Fix integer overflow in GKI timer calculation
Fix an integer overflow in the GKI timer elapsed time calculation that
would cause Bluetooth Stereo Audio (A2DP) playback to stop suddenly
after a period of INT32_MAX microseconds (about 35 minutes) of
continuous uptime of the phone/tablet.

Bug: 16412377

+++ btif/src/btif_media_task.c
@@ -368,7 +368,7 @@ static UINT64 time_now_us()
 {
     struct timespec ts_now;
     clock_gettime(CLOCK_BOOTTIME, &ts_now);
-    return (ts_now.tv_sec * USEC_PER_SEC) + (ts_now.tv_nsec / 1000);
+    return ((UINT64)ts_now.tv_sec * USEC_PER_SEC) + ((UINT64)ts_now.tv_nsec / 1000);
 }
 
 static void log_tstamps_us(char *comment)
